Refactor #68 - moeve `SelectionTable` package util to top
